Move traces to middlewares
